pyproject.toml
[project]
name = "llm-council"
version = "0.2.0"
description = "Your LLM Council - Powered by GitHub Copilot"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "github-copilot-sdk>=0.1.0",
]
